requirements.txt
Flask==1.1.2
Flask-JWT-Extended==3.25.0
Flask-RESTful==0.3.8
matplotlib==3.4.2
pycodestyle==2.6.0
python-dateutil==2.8.1
scipy==1.6.3
Werkzeug==1.0.1
